requirements/base.txt
Django==2.0.1
pytz==2017.3
django-environ==0.4.4
six==1.11.0
djangorestframework==3.7.7
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2017.3
six==1.11.0
django-extensions==1.9.9
django-allauth==0.35
django-crispy-forms==1.7.0
